requirements.txt
Django==1.7.1
django-bootstrap-form==3.1
django-debug-toolbar==1.2.2
django-suit==0.2.12
django-user-accounts==1.0
pinax-theme-bootstrap==5.6.0
Pillow==2.6.1
django-uuid-upload-path
